pyproject.toml
[tool.poetry]
name = "rapidkit-community-staging"
version = "0.1.0"
description = "Public staging playground where we polish the next wave of RapidKit magic together"
authors = ["RapidKit Team <team@rapidkit.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/getrapidkit/community-staging"
repository = "https://github.com/getrapidkit/community-staging"
keywords = ["scaffolding", "boilerplate", "fastapi", "codegen", "devtools"]
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: MIT",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Code Generators",
  "Development Status :: 4 - Beta"
]

packages = [
  { include = "core", from = "src" },
  { include = "cli", from = "src" },
  { include = "kits", from = "src" },
  { include = "modules", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
# Align CLI runtime with engine defaults for compatibility
typer = ">=0.19,<0.20"
click = ">=8.3,<8.4"
rich = "^14.2.0"
pyyaml = "^6.0.3"
tabulate = "^0.9.0"
jsonschema = "^4.25.1"
jinja2 = "^3.1.6"
pydantic = ">=2.12.2"
pydantic-settings = ">=2.11,<3.0"
pynacl = "^1.5.0"
cryptography = "^46.0.2"
types-requests = "^2.32.4.20250913"
requests = "^2.32.5"
email-validator = ">=2.1,<3.0"
pygithub = "^2.7.0"
packaging = "^25.0"
black = "^25.9.0"
yamllint = "^1.37.1"
fastapi = ">=0.119.0"
starlette = ">=0.49.1,<0.50"
uvicorn = { version = ">=0.37.0", extras = ["standard"] }

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-cov = "^7.0.0"
pytest-asyncio = "^0.23.8"
black = "^25.9.0"
ruff = "^0.14.0"
mypy = "^1.18.2"
types-PyYAML = "^6.0.12.20250915"
pip-audit = "^2.9.0"
bandit = "^1.8.6"
pre-commit = "^4.3.0"
types-requests = "^2.32.4.20250913"
cryptography = "^46.0.2"
httpx = "^0.28.0"

[tool.poetry.scripts]
rapidkit = "cli.main:main"

[tool.coverage.run]
branch = false
source = ["src"]
omit = [
  "src/core/quantum_engine.py",
  "src/experimental/*",
  "*/tests/*",
  "*/test_*",
  "*/conftest.py",
  "*/__pycache__/*",
  "*/migrations/*",
  "*/venv/*",
  "*/env/*",
  "*/.venv/*",
  "*/.env/*",
    "*/templates/*",
    "*/pytest-of-*/*",
]

[tool.coverage.report]
skip_empty = true
show_missing = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.excluded_files]
patterns = [
    "*/templates/*",
    "*/migrations/*",
    "*_pb2.py",
    "*/experimental/*",
]
[tool.mypy]
python_version = "3.10"
mypy_path = "src"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
show_error_codes = true
pretty = true
packages = ["core", "cli", "kits", "modules"]
explicit_package_bases = true
namespace_packages = true
follow_imports = "normal"
exclude = [
    "tests/",
    "user-tests/",
    "src/experimental/",
]

[[tool.mypy.overrides]]
module = [
    "cli.ui.printer",
    "core.kit_utils",
    "core.services.snippet_injector",
    "cli.commands.add.module",
    "core.tui",
    "requests.*",
    "jsonschema.*",
    "yaml.*"
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
  "psutil.*",
  "nacl.*"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
